Refactor: make components folder
